Refactor: 컨텐츠 영역 코드 분리 및 툴바 제거 #59
